pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdfstract"
version = "1.0.4"
description = "PDFStract - Unified PDF Extraction & Conversion CLI + Web UI with 10+ extraction libraries"
readme = "CLI_README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "click>=8.1.0",
    "docling>=2.41.0",
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "marker-pdf>=1.8.1",
    "markitdown[pdf]>=0.1.2",
    "paddleocr[all]>=3.3.2",
    "paddlepaddle>=3.2.2; platform_machine != 'aarch64'",
    "pdf2image>=1.17.0",
    "pillow>=10.4.0",
    "pymupdf4llm>=0.0.26",
    "pypdf2>=3.0.1",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.20",
    "rich>=13.0.0",
    "uvicorn[standard]>=0.35.0",
    # Logging
    "loguru>=0.7.2",
    # DeepSeek-OCR dependencies (transformers-based, works on CPU or GPU)
    "transformers>=4.51.1",
    "torch>=2.0.0",
    "addict>=2.4.0",
    "matplotlib>=3.10.7",
    "easydict>=1.13",
    # For Tesseract OCR (requires system tesseract binary: e.g. brew/apt install tesseract)
    "pytesseract>=0.3.10",
    # Unstructured.io for document extraction
    "unstructured[pdf]>=0.15.0",
    # Text chunking for RAG/embedding pipelines (all chunkers enabled)
    "chonkie[all]>=1.0.0",  # Includes semantic, code, neural, genie, and all other chunkers
]

[project.scripts]
pdfstract = "cli:pdfstract"

[project.optional-dependencies]

[tool.setuptools]
py-modules = ["cli", "batch_scheduler"]
packages = ["services", "services.converters", "services.chunkers"]
include-package-data = true
